Removes Travis configuration file & deploy script
We are moving builds of this repo to Cloud Build and no longer need the
Travis config. Additionally, we no longer need the deployment script,
since that step was migrated to the prometheus-support build. All that
needs to be done now is for Cloud Build to run unit tests and test
building the Docker image.