Adds mutex lock to mainteanancestate, removes from handler
The lock in the handler was ultimately to protect from concurrent reads/writes
of state from incoming webhooks. With this commit, a new mutex lock was added
to the maintenancestate struct and was removed from the handler.

Additionally, this commit refactors the Prune() function. Previously it was
making calls to UpdateSite() and UpdateMachine(), failing to consider that
those functions will refuse to do anything for sites that don't exist in
siteinfo, which defeats the whole purpose!