Remove state for retired sites (#46)
* Adds new function to remove retired sites from state

We don't want or need to keep active maintenance states for retired sites. This
commit adds a new maintenancestate.Prune() function which gets run every time
the sites.CachingClient.Reload() gets called to update siteinfo data. It should
remove any sites and machines from maintenance for sites that no longer exist
in siteinfo.

* Replaces all uses of deprecated io/ioutil package

* Moves call to state.Prune() to correct location

* Adds mutex lock to mainteanancestate, removes from handler

The lock in the handler was ultimately to protect from concurrent reads/writes
of state from incoming webhooks. With this commit, a new mutex lock was added
to the maintenancestate struct and was removed from the handler.

Additionally, this commit refactors the Prune() function. Previously it was
making calls to UpdateSite() and UpdateMachine(), failing to consider that
those functions will refuse to do anything for sites that don't exist in
siteinfo, which defeats the whole purpose!

* Updates unit testing to test new Prune() method

* Factors duplicated loop into a function

Previously, the exact same loop was implemented twice in a row. This commit
factors the loop out into a function and makes two calls to this function.

* Prunes state on startup

Previously, state would only get pruned when siteinfo was updated. This small
changes causes GMX to prune the state on startup as well.

* Creates new function removeRetired()

Breaking the actual removal logic from Prune() allow the code to properly defer
the unlocking of the mutex.